requirements.txt
Flask==3.1.2
pymongo==4.15.1
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
